+++ src/bract/cli/main.clj
@@ -43,6 +43,12 @@
 (defn -main
   "This function becomes the Java main() method entry point."
   [& args]
-  (trigger {(key core-kdef/ctx-context-file) "bract-context.edn"
-            (key core-kdef/ctx-inducers)     []
-            (key core-kdef/ctx-cli-args)     (vec args)}))
+  (try
+    (trigger {(key core-kdef/ctx-context-file) "bract-context.edn"
+             (key core-kdef/ctx-inducers)     []
+             (key core-kdef/ctx-cli-args)     (vec args)})
+    (catch Throwable e
+      (when (or (Thread/getDefaultUncaughtExceptionHandler)
+              (.getUncaughtExceptionHandler ^Thread (Thread/currentThread)))
+        (.printStackTrace e))
+      (throw e))))