+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 - [TODO] Use bract.core 0.4.0-alpha2
 - Merge context from `bract-context.edn` in CLI entry point when available
-- [TODO] Catch exception and print stack trace in the main entry point (due to global uncaught handler)
+- Catch exception and print stack trace in the main entry point (due to uncaught handler)
 
 
 ## 0.4.0-alpha1 / 2017-July-31